policer: revamp the api to make it easier to use
This patch adapts the API to make it more MCP oriented and removes the
JWT part. It happens that it would be well more served by a Policer so
it can handle everything regarding authn/authz versus having a split
brain issue with JWT validation on one side and authz on a policer.

Also adds an example policer.