new: support for sbom generation and integrity check
This patch adds the scan subcommand:

- `scan tools`: dump all the tools of a server
- `scan sbom`: generate a sbom file with hashes for integrity check
- `scan check`: verifies the sbom matches am existing sbom file offline.

For example, to dump all the tools available in
modelcontextprotocol/server-everything:

    minibridge scan tools -- \
	npx @modelcontextprotocol/server-everything

And to generate the SBOM file:

    minibridge scan sbom -- \
	npx @modelcontextprotocol/server-everything > server.sbom

To check an sbom offline:

    minibridge scan check server.sbom -- \
	npx @modelcontextprotocol/server-everything

The SBOM file can also be passed to the minibridge backend to dynamicaly
make sure no parts interpreted by the LLM have changed since the SBOM
file generation.

For instance, This will make sure a legit tool that ask
to read a file does not suddenly change to a tool that sends all your
credentials to some shady place.

Example:

    minibridge backend -l :8000 --sbom server.sbom -- \
	npx @modelcontextprotocol/server-everything > server.sbom