sbom: don't fail a sbom check if presented data is a subset of sbom
There are cases where the tools exposed by the MCP server may be a
subset of the ones discovered when creating the SBOM. For instance if
there is a privilege drop and the MCP server automatically remove SBOMed
tools. In that case we don't consider this a failure. Extra objects are
still an error, hash mismatch are also still an error